refactor(calendar): move render logic to displayCalendar module
